+++ R/plot-volcano.R
@@ -1,6 +1,6 @@
 #' Create a volcano plot from expression data
 #'
-#' Create a volcano plot from the given dataframe
+#' Create a volcano plot from a data.frame containing differential expression results.
 #' @param df dataframe containing columns with gene names, p-values, and log-fold changes.
 #' @param x logFC column. Default (logFC)
 #' @param y FDR or p-value column. Default (FDR)
@@ -9,12 +9,22 @@
 #' @param lfc log-fold-change cutoff for plotting. Values greater than the abs(lfc) and less than fdr are displayed as differentially expressed. Default (0)
 #' @param label_sig TRUE/FALSE. apply ggrepel::geom_text_labels to significant DE genes.
 #' @param annotate_counts TRUE/FALSE. Annotate the plot with the summarized gene counts
+#' @param up_color Point color of the up-regulated features
+#' @param down_color Point color of the down-regulated features
+#' @param nonde_color Point color of the unperturbed features
+#' @param up_alpha Point alpha of the up-regulated features
+#' @param down_alpha Point alpha of the down-regulated features
+#' @param nonde_alpha Point alpha of the unperturbed features
+#' @param up_size Point size of the up-regulated features
+#' @param down_size Point size of the down-regulated features
+#' @param nonde_size Point size of the unperturbed features
 #' @param xmin_label_offset numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'down' counts
 #' @param xmax_label_offset numeric. Value between 0 and 1 inclusive to control the x-position of the count annotation label for the 'up' counts
 #' @param ymax_label_offset numeric. Value between 0 and 1 inclusive to control the y-position of the count labels.
+#' @param ... Additional arguments passed to \code{ggrepel::geom_text_repel}
 #' @return ggplot volcano plot
+#' @import data.table
 #' @export
-#' @importFrom magrittr %>%
 #' @examples
 #' \dontrun{
 #' library(edgeR)
@@ -47,32 +57,32 @@
 #' plot_volcano(res_df)
 #' }
 #'
-plot_volcano <- function(df,
-                         x = logFC,
-                         y = FDR,
-                         lab = feature_id,
-                         fdr = 0.05,
-                         lfc = 0,
-                         label_sig = FALSE,
-                         annotate_counts = TRUE,
-                         xmin_label_offset = 0.5,
-                         xmax_label_offset = 0.5,
-                         ymax_label_offset = 0.8) {
-  stopifnot("xmin_label_offset must be between 0 and 1" = xmin_label_offset >= 0 & xmin_label_offset <= 1)
-  stopifnot("xmax_label_offset must be between 0 and 1" = xmax_label_offset >= 0 & xmax_label_offset <= 1)
-  stopifnot("ymax_label_offset must be between 0 and 1" = ymax_label_offset >= 0 & ymax_label_offset <= 1)
+plot_volcano <- function(df, x = "logFC", y = "FDR", lab = "feature_id", fdr = 0.1, lfc = 0,
+                         label_sig = FALSE, annotate_counts = TRUE, up_color = "red2", down_color = "red2",
+                         nonde_color = "grey40", up_alpha = 1, down_alpha = 1, nonde_alpha = 1, up_size = 1,
+                         down_size = 1, nonde_size = 1, xmin_label_offset = 0.5,
+                         xmax_label_offset = 0.5, ymax_label_offset = 0.8, ...) {
+  dt <- as.data.table(df)
 
-  plot_df <- df %>%
-    dplyr::mutate(signif = dplyr::if_else({{ y }} < fdr & abs({{ x }}) > lfc, "yes", "no"))
+  # Modify the input df for plotting
+  dt[, `:=`(
+    direction = fcase(
+      get(y) < fdr & abs(get(x)) > lfc & get(x) > 0, "Up",
+      get(y) < fdr & abs(get(x)) > lfc & get(x) < 0, "Down",
+      default = "Unperturbed"
+    ),
+    logPval = -log10(get(y))
+  )]
 
-  vplot <- ggplot2::ggplot(data = plot_df, ggplot2::aes(x = {{ x }}, y = -log10({{ y }}), label = {{ lab }})) +
-    ggplot2::geom_point(ggplot2::aes(color = .data$signif)) +
-    ggplot2::scale_colour_manual(values = c("no" = "gray40", "yes" = "red2")) +
+  p <- ggplot2::ggplot(data = dt, ggplot2::aes_string(x = x, y = "logPval", label = lab)) +
+    ggplot2::geom_point(data = dt[direction == "Unperturbed"], color = nonde_color, alpha = nonde_alpha, size = nonde_size) +
+    ggplot2::geom_point(data = dt[direction == "Down"], color = down_color, alpha = down_alpha, size = down_size) +
+    ggplot2::geom_point(data = dt[direction == "Up"], color = up_color, alpha = up_alpha, size = up_size) +
     ggplot2::geom_vline(xintercept = 0, linetype = 2) +
     ggplot2::geom_vline(xintercept = lfc, linetype = 3) +
     ggplot2::geom_vline(xintercept = -lfc, linetype = 3) +
     ggplot2::geom_hline(yintercept = -log10(fdr), linetype = 3) +
-    ggplot2::labs(color = "Significant") +
+    ggplot2::labs(color = "Direction of Expression") +
     ggplot2::labs(
       caption = paste("FDR = ", fdr, "\nlfc cutoff = ", round(lfc, digits = 2)),
       x = "logFC",
@@ -83,23 +93,24 @@ plot_volcano <- function(df,
 
   # add text labels to significant genes
   if (label_sig) {
-    vplot <- vplot +
+    p <- p +
       ggrepel::geom_text_repel(
-        data = plot_df %>% dplyr::filter(signif == "yes"),
-        ggplot2::aes(label = {{ lab }})
+        data = dt[direction %in% c("Up", "Down")],
+        ggplot2::aes_string(label = lab),
+        ...
       )
   }
 
   if (annotate_counts) {
-    d <- coriell::summarize_dge(df, fdr_col = {{ y }}, lfc_col = {{ x }}, fdr = fdr, lfc = lfc)
-    plot_lims <- coriell::get_axis_limits(vplot)
+    d <- coriell::summarize_dge(df, fdr_col = y, lfc_col = x, fdr = fdr, lfc = lfc)
+    plot_lims <- coriell::get_axis_limits(p)
 
-    up_count <- d[d$dge == "up", "n", drop = TRUE]
-    down_count <- d[d$dge == "down", "n", drop = TRUE]
-    up_pct <- round(d[d$dge == "up", "perc", drop = TRUE], digits = 2)
-    down_pct <- round(d[d$dge == "down", "perc", drop = TRUE], digits = 2)
+    up_count <- d[d$Direction == "Up", "N", drop = TRUE]
+    down_count <- d[d$Direction == "Down", "N", drop = TRUE]
+    up_pct <- round(d[d$Direction == "Up", "Percent", drop = TRUE], digits = 2)
+    down_pct <- round(d[d$Direction == "Down", "Percent", drop = TRUE], digits = 2)
 
-    vplot <- vplot +
+    p <- p +
       ggplot2::annotate(
         geom = "label",
         x = xmin_label_offset * plot_lims$x_min,
@@ -113,5 +124,5 @@ plot_volcano <- function(df,
         label = paste0(up_count, "\n", up_pct, "%")
       )
   }
-  vplot
+  p
 }